moving include file to includes folder
for courses and projects model
