+++ pages/Courses_Projects/delete_course.php
@@ -12,7 +12,7 @@
  */
 
 
-require_once ("CoursesProjectsModel.php");
+require_once ("../../includes/courses_projects_model.php");
 require_once ("../../includes/session.php");
 require_once("../../includes/db_connection.php");
 require_once("../../includes/functions.php");
