+++ pages/Courses_Projects/update_project.php
@@ -19,7 +19,7 @@
  * @param integer tag       teg of the project to be updated
  */
 
-require_once ("CoursesProjectsModel.php");
+require_once ("../../includes/courses_projects_model.php");
 require_once ("../../includes/session.php");
 require_once("../../includes/db_connection.php");
 require_once("../../includes/functions.php");
